greenhouse, rule chain (47)

+++ src/main/resources/files/Greenhouse/finish_energy_meter.js
@@ -1,11 +1,11 @@
 var newMsg = {
     'ts' : parseInt(metadata.ts),
     'values' : {
-        'energyConsumptionLight' : parseInt(metadata.values_energyConsumptionLight),
-        'energyConsumptionHeating' : parseInt(metadata.values_energyConsumptionHeating),
-        'energyConsumptionCooling' : parseInt(metadata.values_energyConsumptionCooling),
-        'energyConsumptionAirControl' : parseInt(metadata.values_energyConsumptionAirControl),
-        'energyConsumptionIrrigation' : parseInt(metadata.values_energyConsumptionIrrigation)
+        'energyConsumptionLight' : parseFloat(metadata.values_energyConsumptionLight),
+        'energyConsumptionHeating' : parseFloat(metadata.values_energyConsumptionHeating),
+        'energyConsumptionCooling' : parseFloat(metadata.values_energyConsumptionCooling),
+        'energyConsumptionAirControl' : parseFloat(metadata.values_energyConsumptionAirControl),
+        'energyConsumptionIrrigation' : parseFloat(metadata.values_energyConsumptionIrrigation)
     }
 };
 return {msg: newMsg, metadata: metadata, msgType: msgType};
