+++ src/main/resources/files/Greenhouse/finish_harvest_reporter.js
@@ -5,7 +5,7 @@ if (iteratedDate.getHours() > 0) {
 var newMsg = {
     'ts' : parseInt(metadata.ts),
     'values' : {
-        'cropWeight' : parseInt(metadata.values_cropWeight),
+        'cropWeight' : parseFloat(metadata.values_cropWeight),
         'workerInCharge' : metadata.values_workerInCharge
     }
 };
